requirements.txt
Django==3.2.2
pandas==1.2.4
scikit-learn==0.24.2
scikit-surprise==1.1.1
